extract GetStatusFile to GetJSONFile
