+++ utils/sys_utils_test.go
@@ -2,6 +2,7 @@ package utils_test
 
 import (
 	"errors"
+	"io/ioutil"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -11,6 +12,53 @@ import (
 	"github.com/greenplum-db/gpupgrade/upgrade"
 	"github.com/greenplum-db/gpupgrade/utils"
 )
+
+func TestJSONFile(t *testing.T) {
+	stateDir, err := ioutil.TempDir("", "")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer func() {
+		if err := os.RemoveAll(stateDir); err != nil {
+			t.Errorf("removing temp directory: %v", err)
+		}
+	}()
+
+	JSONFileName := "file.json"
+	path := filepath.Join(stateDir, JSONFileName)
+
+	t.Run("creates JSON file if it does not exist", func(t *testing.T) {
+		_, err := os.Open(path)
+		if !os.IsNotExist(err) {
+			t.Errorf("returned error %#v want ErrNotExist", err)
+		}
+
+		statusFile, err := utils.GetJSONFile(stateDir, JSONFileName)
+		if err != nil {
+			t.Errorf("unexpected error %v", err)
+		}
+
+		contents := testutils.MustReadFile(t, statusFile)
+		if contents != "{}" {
+			t.Errorf("read %q want {}", contents)
+		}
+	})
+
+	t.Run("does not create JSON file if it already exists", func(t *testing.T) {
+		expected := "1234"
+		testutils.MustWriteToFile(t, path, expected)
+
+		statusFile, err := utils.GetJSONFile(stateDir, JSONFileName)
+		if err != nil {
+			t.Errorf("unexpected error %v", err)
+		}
+
+		contents := testutils.MustReadFile(t, statusFile)
+		if contents != expected {
+			t.Errorf("read %q want %q", contents, expected)
+		}
+	})
+}
 
 func TestMove(t *testing.T) {
 	t.Run("move run successfully", func(t *testing.T) {
